fix(mobile): android push notification redirection
